State: Move isRequestingJetpackConnectionStatus selector to global selectors directory
