+++ client/state/selectors/index.js
@@ -37,6 +37,7 @@ export isDomainOnlySite from './is-domain-only-site';
 export isPrivateSite from './is-private-site';
 export isPublicizeEnabled from './is-publicize-enabled';
 export isRequestingBillingTransactions from './is-requesting-billing-transactions';
+export isRequestingJetpackConnectionStatus from './is-requesting-jetpack-connection-status';
 export isRequestingPostLikes from './is-requesting-post-likes';
 export isRequestingTimezones from './is-requesting-timezones';
 export isRequestingSharingButtons from './is-requesting-sharing-buttons';
