+++ client/state/selectors/test/is-requesting-jetpack-connection-status.js
@@ -1,0 +1,51 @@
+/**
+ * External dependencies
+ */
+import { expect } from 'chai';
+
+/**
+ * Internal dependencies
+ */
+import { isRequestingJetpackConnectionStatus } from '../';
+import { requests as REQUESTS_FIXTURE } from './fixtures/jetpack-connection';
+
+describe( 'isRequestingJetpackConnectionStatus()', () => {
+	it( 'should return true if the connection status is being fetched', () => {
+		const stateIn = {
+				jetpack: {
+					jetpackConnection: {
+						requests: REQUESTS_FIXTURE
+					}
+				}
+			},
+			siteId = 87654321;
+		const output = isRequestingJetpackConnectionStatus( stateIn, siteId );
+		expect( output ).to.be.true;
+	} );
+
+	it( 'should return false if the connection status is not being fetched', () => {
+		const stateIn = {
+				jetpack: {
+					jetpackConnection: {
+						requests: REQUESTS_FIXTURE
+					}
+				}
+			},
+			siteId = 12345678;
+		const output = isRequestingJetpackConnectionStatus( stateIn, siteId );
+		expect( output ).to.be.false;
+	} );
+
+	it( 'should return null if the site is not known yet', () => {
+		const stateIn = {
+				jetpack: {
+					jetpackConnection: {
+						requests: REQUESTS_FIXTURE
+					}
+				}
+			},
+			siteId = 88888888;
+		const output = isRequestingJetpackConnectionStatus( stateIn, siteId );
+		expect( output ).to.be.null;
+	} );
+} );
